requirements.txt
asgiref==3.3.1
Django==3.1.5
django-crispy-forms==1.10.0
django-stronghold==0.4.0
mysqlclient==2.0.3
python-dotenv==0.16.0
pytz==2020.5
sqlparse==0.4.1
